feat: 추가 자녀성장일지 기능

+++ UKIDS/src/main/java/com/modernfamily/ukids/domain/growthFolder/model/service/GrowthFolderServiceImpl.java
@@ -2,6 +2,9 @@ package com.modernfamily.ukids.domain.growthFolder.model.service;
 
 import com.modernfamily.ukids.domain.family.entity.Family;
 import com.modernfamily.ukids.domain.family.model.repository.FamilyRepository;
+import com.modernfamily.ukids.domain.familyMember.entity.FamilyMember;
+import com.modernfamily.ukids.domain.familyMember.model.repository.FamilyMemberRepository;
+import com.modernfamily.ukids.domain.growthFolder.dto.GrowthFolderPaginationDto;
 import com.modernfamily.ukids.domain.growthFolder.dto.GrowthFolderRequestDto;
 import com.modernfamily.ukids.domain.growthFolder.dto.GrowthFolderResponseDto;
 import com.modernfamily.ukids.domain.growthFolder.dto.GrowthFolderUpdateRequestDto;
@@ -12,8 +15,10 @@ import com.modernfamily.ukids.domain.user.dto.CustomUserDetails;
 import com.modernfamily.ukids.global.exception.CustomException;
 import com.modernfamily.ukids.global.exception.ExceptionResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.*;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Repository
@@ -22,45 +27,65 @@ public class GrowthFolderServiceImpl implements GrowthFolderService {
     private final GrowthFolderRepository growthFolderRepository;
     private final FamilyRepository familyRepository;
     private final GrowthFolderMapper growthFolderMapper;
+    private final FamilyMemberRepository familyMemberRepository;
+
     @Override
-    public GrowthFolderResponseDto createGrowthFolder(GrowthFolderRequestDto growthFolderRequestDto) {
+    public void createGrowthFolder(GrowthFolderRequestDto growthFolderRequestDto) {
         Family family = familyRepository.findByFamilyId(growthFolderRequestDto.getFamilyId())
                 .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_FAMILY_EXCEPTION));
 
         GrowthFolder growthFolder = growthFolderRepository.save(growthFolderMapper.toGrowthFolderRequestEntity(growthFolderRequestDto));
 
 
-        return growthFolderMapper.toGrowthFolderResponseDto(growthFolder);
+        growthFolderMapper.toGrowthFolderResponseDto(growthFolder);
     }
 
     @Override
-    public List<GrowthFolderResponseDto> getGrowthFolders(Long familyId) {
+    public GrowthFolderPaginationDto getGrowthFolders(Long familyId, int size, int page) {
         Family family = familyRepository.findByFamilyId(familyId)
                 .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_FAMILY_EXCEPTION));
 
-        List<GrowthFolder> growthFolders = growthFolderRepository.getGrowthFolders(familyId);
 
-        if(growthFolders == null || growthFolders.isEmpty()){
+        Pageable pageable = PageRequest.of(page-1, size, Sort.by(Sort.Direction.DESC, "createTime"));
+        Page<GrowthFolder> growthFolderPage = growthFolderRepository.findAllByFamily_FamilyId(familyId, pageable);
+        if(!growthFolderPage.hasContent()) {
             throw new ExceptionResponse(CustomException.NOT_FOUND_GROWTHFOLDER_EXCEPTION);
         }
 
-        // 추후 FamilyMember에 userId와 familyId를 이용해 승인된 유저인지 조회
+        // FamilyMember에 승인된 인원인지 조회
+        String id = CustomUserDetails.contextGetUserId();
+        FamilyMember familyMember = familyMemberRepository.findByUser_IdAndFamily_FamilyId(id, familyId)
+                .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_FAMILYMEMBER_EXCEPTION));
 
-        return growthFolderMapper.toGrowthFolderResponseDtoList(growthFolders);
+        if(!familyMember.isApproval())
+            throw new ExceptionResponse(CustomException.NOT_APPROVAL_FAMILYMEMBER_EXCEPTION);
+
+        List<GrowthFolderResponseDto> growthFolders = new ArrayList<>();
+
+        for (GrowthFolder growthFolder : growthFolderPage) {
+            growthFolders.add(growthFolderMapper.toGrowthFolderResponseDto(growthFolder));
+        }
+        return new GrowthFolderPaginationDto(growthFolders, growthFolderPage.getTotalPages(), size, page);
+
     }
 
     @Override
-    public GrowthFolderResponseDto updateGrowthFolder(GrowthFolderUpdateRequestDto growthFolderUpdateRequestDto) {
-        // 추후 FamilyMember에 userId와 familyId를 이용해 승인된 유저인지 조회
+    public void updateGrowthFolder(GrowthFolderUpdateRequestDto growthFolderUpdateRequestDto) {
 
         // 폴더가 있는지 확인
         GrowthFolder growthFolder = growthFolderRepository.findByFolderId(growthFolderUpdateRequestDto.getFolderId())
                         .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_GROWTHFOLDER_EXCEPTION));
 
-        growthFolderRepository.updateGrowthFolder(growthFolderUpdateRequestDto);
-        growthFolder = growthFolderRepository.findByFolderId(growthFolderUpdateRequestDto.getFolderId())
-                .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_GROWTHFOLDER_EXCEPTION));
-        return growthFolderMapper.toGrowthFolderResponseDto(growthFolder);
+        // FamilyMember에 승인된 유저인지 조회
+        String id = CustomUserDetails.contextGetUserId();
+        FamilyMember familyMember = familyMemberRepository.findByUser_IdAndFamily_FamilyId(id, growthFolder.getFamily().getFamilyId())
+                .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_FAMILYMEMBER_EXCEPTION));
+
+        if(!familyMember.isApproval())
+            throw new ExceptionResponse(CustomException.NOT_APPROVAL_FAMILYMEMBER_EXCEPTION);
+
+
+        growthFolderRepository.updateGrowthFolder(growthFolderMapper.toGrowthFolderUpdateEntity(growthFolderUpdateRequestDto));
     }
 
     @Override
@@ -69,8 +94,14 @@ public class GrowthFolderServiceImpl implements GrowthFolderService {
         GrowthFolder growthFolder = growthFolderRepository.findByFolderId(folderId)
                 .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_GROWTHFOLDER_EXCEPTION));
 
+        // FamilyMember에 승인된 유저인지 조회
+        String id = CustomUserDetails.contextGetUserId();
+        Family family = growthFolder.getFamily();
+        FamilyMember familyMember = familyMemberRepository.findByUser_IdAndFamily_FamilyId(id, family.getFamilyId())
+                .orElseThrow(() -> new ExceptionResponse(CustomException.NOT_FOUND_FAMILYMEMBER_EXCEPTION));
 
-        // userId와 familyId를 이용해 승인된 유저인지 조회해야함
+        if(!familyMember.isApproval())
+            throw new ExceptionResponse(CustomException.NOT_APPROVAL_FAMILYMEMBER_EXCEPTION);
 
         // 삭제 update
         growthFolderRepository.deleteGrowthFolder(folderId);
